update the app to mp4 and use the best practices
