--- finish/barista/src/main/java/com/sebastian_daschner/barista/boundary/HealthResource.java
@@ -1,28 +1,0 @@
-package com.sebastian_daschner.barista.boundary;
-
-import javax.enterprise.context.ApplicationScoped;
-import org.eclipse.microprofile.health.Health;
-import org.eclipse.microprofile.health.HealthCheck;
-import org.eclipse.microprofile.health.HealthCheckResponse;
-
-@Health
-@ApplicationScoped
-public class HealthResource implements HealthCheck {
-	
-  public boolean isHealthy() {
-      // Nothing to check in this implementation
-      return true;
-  }
-
-  @Override
-  public HealthCheckResponse call() {
-    if (!isHealthy()) {
-      return HealthCheckResponse.named(this.getClass().getSimpleName())
-                                .down()
-                                .build();
-    }
-    return HealthCheckResponse.named(this.getClass().getSimpleName())
-                              .up().build();
-  }
-
-}
