+++ finish/barista/src/main/java/com/sebastian_daschner/barista/entity/CoffeeBrew.java
@@ -1,7 +1,5 @@
 package com.sebastian_daschner.barista.entity;
 
-import javax.json.bind.annotation.JsonbProperty;
-
 public class CoffeeBrew {
 
     private CoffeeType type;
