+++ finish/barista/src/main/liberty/config/server.xml
@@ -4,10 +4,15 @@
     <featureManager>
         <feature>jaxrs-2.1</feature>
         <feature>cdi-2.0</feature>
-        <feature>mpOpenAPI-1.1</feature>
-        <feature>mpHealth-2.2</feature>
+        <feature>mpOpenAPI-2.0</feature>
+        <feature>mpHealth-3.0</feature>
     </featureManager>
 
-    <httpEndpoint host="*" httpPort="9081" httpsPort="9444" id="defaultHttpEndpoint"/>
+    <variable name="default.http.port" defaultValue="9081"/>
+    <variable name="default.https.port" defaultValue="9444"/>
+  
+    <httpEndpoint id="defaultHttpEndpoint" host="*" 
+        httpPort="${default.http.port}" 
+        httpsPort="${default.https.port}"/>
 
 </server>
