+++ finish/coffee-shop/configDropins/overrides/metrics-prod.xml
@@ -12,9 +12,7 @@
      for demo purposes only and MUST NOT BE USED IN PRODUCTION
      AS IT IS INSECURE. -->  
     <variable name="admin.password" value="change_it" />
-    <variable name="keystore.password" value="change_it" />
     
     <quickStartSecurity userName="admin" userPassword="${admin.password}"/>
-    <keyStore id="defaultKeyStore" password="${keystore.password}"/>    
      
 </server>
