+++ finish/coffee-shop/pom.xml
@@ -6,12 +6,6 @@
     <version>1.0-SNAPSHOT</version>
     <packaging>war</packaging>
 
-    <parent>
-        <groupId>net.wasdev.wlp.maven.parent</groupId>
-        <artifactId>liberty-maven-app-parent</artifactId>
-        <version>2.6.2</version>
-    </parent>
-
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
@@ -19,21 +13,13 @@
         <maven.compiler.target>1.8</maven.compiler.target>
 
         <!-- Liberty Server Port Configuration-->
-        <testServerHttpPort>9080</testServerHttpPort>
-        <testServerHttpsPort>9443</testServerHttpsPort>
+        <liberty.var.default.http.port>9080</liberty.var.default.http.port>
+        <liberty.var.default.https.port>9443</liberty.var.default.https.port>
         <baristaBaseURL>http://localhost:9081</baristaBaseURL>
-        
-        <failOnMissingWebXml>false</failOnMissingWebXml>
-
-        <!-- Plugin Versions-->
-        <version.liberty-maven-plugin>3.3.4</version.liberty-maven-plugin>
-        <version.maven-compiler-plugin>3.5.1</version.maven-compiler-plugin>
-        <version.maven-failsafe-plugin>3.0.0-M4</version.maven-failsafe-plugin>
-        <version.maven-war-plugin>3.2.3</version.maven-war-plugin>
     </properties>
 
     <dependencies>
-      <!--Open Liberty features -->
+        <!--Open Liberty provided features -->
         <dependency>
             <groupId>jakarta.platform</groupId>
             <artifactId>jakarta.jakartaee-web-api</artifactId>
@@ -43,46 +29,35 @@
         <dependency>
             <groupId>org.eclipse.microprofile</groupId>
             <artifactId>microprofile</artifactId>
-            <version>3.3</version>
+            <version>4.0.1</version>
             <type>pom</type>
             <scope>provided</scope>
         </dependency> 
     </dependencies>
 
     <build>
-        <finalName>coffee-shop</finalName>
+        <finalName>${project.artifactId}</finalName>
 
         <plugins>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-compiler-plugin</artifactId>
-                <version>${version.maven-compiler-plugin}</version>
-                <configuration>
-                    <source>1.8</source>
-                    <target>1.8</target>
-                </configuration>
-            </plugin>
-
-            <plugin>
                 <artifactId>maven-war-plugin</artifactId>
-                <version>${version.maven-war-plugin}</version>
+                <version>3.3.1</version>
             </plugin>
-
             <plugin>
                 <groupId>io.openliberty.tools</groupId>
                 <artifactId>liberty-maven-plugin</artifactId>
-                <version>${version.liberty-maven-plugin}</version>
-                <configuration>
-                    <include>${packaging.type}</include>
-                    <packageFile>${project.build.directory}/${package.file}</packageFile>
-                    <bootstrapProperties>
-                        <war.name>${artifactId}.war</war.name>
-                        <env.default_http_port>${testServerHttpPort}</env.default_http_port>
-                        <env.default_https_port>${testServerHttpsPort}</env.default_https_port>
-                        <default_barista_base_url>${baristaBaseURL}</default_barista_base_url>
-                    </bootstrapProperties>
-                    <looseApplication>false</looseApplication>
-                </configuration>
+                <version>3.3.4</version>
+            </plugin>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-failsafe-plugin</artifactId>
+                <version>2.22.2</version>
+            </plugin>   
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-surefire-plugin</artifactId>
+                <version>2.22.2</version>
             </plugin>
         </plugins>
     </build>
