+++ finish/coffee-shop/src/main/webapp/META-INF/microprofile-config.properties
@@ -1,0 +1,1 @@
+default_barista_base_url=http://localhost:9081
