+++ start/barista/src/main/liberty/config/server.xml
@@ -4,10 +4,12 @@
     <featureManager>
         <feature>jaxrs-2.1</feature>
         <feature>cdi-2.0</feature>
-        <feature>mpOpenAPI-1.1</feature>
-        <feature>mpHealth-2.2</feature>
+        <feature>mpOpenAPI-2.0</feature>
+        <feature>mpHealth-3.0</feature>
     </featureManager>
 
-    <httpEndpoint host="*" httpPort="${env.default.http.port}" httpsPort="${env.default.https.port}" id="defaultHttpEndpoint"/>
+    <httpEndpoint id="defaultHttpEndpoint" host="*" 
+        httpPort="9081" 
+        httpsPort="9444"/>
 
 </server>
