--- start/coffee-shop/configDropins/overrides/metrics-prod.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<server description="Coffee Shop Server">
-
-    <featureManager>
-        <feature>mpMetrics-2.3</feature>
-    </featureManager>
-    
-    <mpMetrics authentication="true" />
-
-     <!-- 
-     Note, this configuration is
-     for demo purposes only and MUST NOT BE USED IN PRODUCTION
-     AS IT IS INSECURE. -->  
-    <variable name="admin.password" value="change_it" />
-    <variable name="keystore.password" value="change_it" />
-    
-    <quickStartSecurity userName="admin" userPassword="${admin.password}"/>
-    <keyStore id="defaultKeyStore" password="${keystore.password}"/>    
-     
-</server>
