+++ start/coffee-shop/src/main/java/com/sebastian_daschner/coffee_shop/entity/CoffeeBrew.java
@@ -1,7 +1,5 @@
 package com.sebastian_daschner.coffee_shop.entity;
 
-import com.sebastian_daschner.coffee_shop.entity.CoffeeType;
-
 public class CoffeeBrew {
 
     private CoffeeType type;
